refactor(app): decompose god files into focused modules
Extract pure utilities and composables from three oversized files:
- useMetadata.ts (2452→2252 lines): idle scheduler, promise store, fetch response types, item picker
- useTaskFiltering.ts (1099→878 lines): status/trader count calculations
- StreamerToolsPanel.vue (1276→861 lines): overlay state and persistence logic
