+++ app/composables/useTaskFiltering.ts
@@ -1,3 +1,4 @@
+import { useTaskCounts } from '@/composables/useTaskCounts';
 import { isMapObjectiveType } from '@/features/tasks/taskObjectiveTypes';
 import { useMetadataStore } from '@/stores/useMetadata';
 import { usePreferencesStore } from '@/stores/usePreferences';
@@ -855,229 +856,7 @@ export function useTaskFiltering() {
       reloadingTasks.value = false;
     }
   };
-  /**
-   * Calculate task counts by status (all, available, locked, completed)
-   */
-  const calculateStatusCounts = (
-    userView: string
-  ): { all: number; available: number; locked: number; completed: number; failed: number } => {
-    const perfTimer = perfStart('[Tasks] calculateStatusCounts', {
-      tasks: metadataStore.tasks.length,
-      userView,
-    });
-    const counts = { all: 0, available: 0, locked: 0, completed: 0, failed: 0 };
-    const taskList = metadataStore.tasks;
-    // Get task type filter settings
-    const showKappa = !preferencesStore.getHideNonKappaTasks;
-    const showLightkeeper = preferencesStore.getShowLightkeeperTasks;
-    const showNonSpecial = preferencesStore.getShowNonSpecialTasks;
-    const hasTypeSelection = showKappa || showLightkeeper || showNonSpecial;
-    const onlyTasksWithRequiredKeys = shouldApplyRequiredKeysFilter();
-    // Get prestige filtering data
-    const userPrestigeLevel = tarkovStore.getPrestigeLevel();
-    const prestigeTaskMap = metadataStore.prestigeTaskMap || new Map<string, number>();
-    // Get edition-based excluded tasks
-    const userEdition = tarkovStore.getGameEdition();
-    const excludedTaskIds = metadataStore.getExcludedTaskIdsForEdition(userEdition);
-    for (const task of taskList) {
-      // Skip tasks not available for user's game edition
-      if (excludedTaskIds.has(task.id)) continue;
-      // Skip prestige tasks that don't match user's prestige level
-      if (prestigeTaskMap.has(task.id)) {
-        const taskPrestigeLevel = prestigeTaskMap.get(task.id);
-        if (taskPrestigeLevel !== userPrestigeLevel) continue;
-      }
-      // Skip tasks filtered out by task type settings
-      const isKappaRequired = task.kappaRequired === true;
-      const isLightkeeperRequired = task.lightkeeperRequired === true;
-      const isNonSpecial = !isKappaRequired && !isLightkeeperRequired;
-      if (hasTypeSelection) {
-        const matchesTaskType =
-          (isKappaRequired && showKappa) ||
-          (isLightkeeperRequired && showLightkeeper) ||
-          (isNonSpecial && showNonSpecial);
-        if (!matchesTaskType) continue;
-      }
-      if (onlyTasksWithRequiredKeys && !taskHasRequiredKeys(task)) {
-        continue;
-      }
-      if (isAllUsersView(userView)) {
-        const teamIds = Object.keys(progressStore.visibleTeamStores || {});
-        const relevantTeamIds = teamIds.filter((teamId) => {
-          const teamFaction = progressStore.playerFaction[teamId];
-          const taskFaction = task.factionName;
-          return taskFaction === 'Any' || taskFaction === teamFaction;
-        });
-        if (relevantTeamIds.length === 0) continue;
-        counts.all++;
-        const isFailedForAny = relevantTeamIds.some(
-          (teamId) => progressStore.tasksFailed?.[task.id]?.[teamId] === true
-        );
-        const isAvailableForAny = relevantTeamIds.some((teamId) => {
-          const isUnlocked = progressStore.unlockedTasks?.[task.id]?.[teamId] === true;
-          const isCompleted = progressStore.tasksCompletions?.[task.id]?.[teamId] === true;
-          const isFailed = progressStore.tasksFailed?.[task.id]?.[teamId] === true;
-          return isUnlocked && !isCompleted && !isFailed;
-        });
-        const isCompletedByAll = relevantTeamIds.every((teamId) => {
-          return (
-            progressStore.tasksCompletions?.[task.id]?.[teamId] === true &&
-            progressStore.tasksFailed?.[task.id]?.[teamId] !== true
-          );
-        });
-        if (isFailedForAny) {
-          counts.failed++;
-        } else if (isCompletedByAll) {
-          counts.completed++;
-        } else if (isAvailableForAny && !isTaskInvalid(task.id, 'all')) {
-          counts.available++;
-        } else if (!isTaskInvalid(task.id, 'all')) {
-          counts.locked++;
-        }
-      } else {
-        // For single user view
-        const taskFaction = task.factionName;
-        const userFaction = progressStore.playerFaction[userView];
-        if (taskFaction !== 'Any' && taskFaction !== userFaction) continue;
-        counts.all++;
-        const isUnlocked = progressStore.unlockedTasks?.[task.id]?.[userView] === true;
-        const isCompleted = progressStore.tasksCompletions?.[task.id]?.[userView] === true;
-        const isFailed = progressStore.tasksFailed?.[task.id]?.[userView] === true;
-        if (isFailed) {
-          counts.failed++;
-        } else if (isCompleted) {
-          counts.completed++;
-        } else if (isUnlocked && !isTaskInvalid(task.id, userView)) {
-          // Only count as available if not permanently invalid/blocked
-          counts.available++;
-        } else if (!isTaskInvalid(task.id, userView)) {
-          // Only count as locked if not permanently invalid/blocked
-          counts.locked++;
-        }
-      }
-    }
-    perfEnd(perfTimer, { total: counts.all });
-    return counts;
-  };
-  /**
-   * Calculate task counts per trader based on current status filter
-   */
-  const calculateTraderCounts = (
-    userView: string,
-    secondaryView: TaskSecondaryView = 'available'
-  ): Record<string, number> => {
-    const perfTimer = perfStart('[Tasks] calculateTraderCounts', {
-      tasks: metadataStore.tasks.length,
-      userView,
-      secondaryView,
-    });
-    const counts: Record<string, number> = {};
-    const taskList = metadataStore.tasks;
-    const showKappa = !preferencesStore.getHideNonKappaTasks;
-    const showLightkeeper = preferencesStore.getShowLightkeeperTasks;
-    const showNonSpecial = preferencesStore.getShowNonSpecialTasks;
-    const hasTypeSelection = showKappa || showLightkeeper || showNonSpecial;
-    const onlyTasksWithRequiredKeys = shouldApplyRequiredKeysFilter();
-    const userPrestigeLevel = tarkovStore.getPrestigeLevel();
-    const prestigeTaskMap = metadataStore.prestigeTaskMap || new Map<string, number>();
-    const userEdition = tarkovStore.getGameEdition();
-    const excludedTaskIds = metadataStore.getExcludedTaskIdsForEdition(userEdition);
-    const isAvailableStatus = (status: {
-      isUnlocked: boolean;
-      isCompleted: boolean;
-      isFailed: boolean;
-    }) => status.isUnlocked && !status.isCompleted && !status.isFailed;
-    for (const task of taskList) {
-      if (excludedTaskIds.has(task.id)) continue;
-      if (prestigeTaskMap.has(task.id)) {
-        const taskPrestigeLevel = prestigeTaskMap.get(task.id);
-        if (taskPrestigeLevel !== userPrestigeLevel) continue;
-      }
-      const isKappaRequired = task.kappaRequired === true;
-      const isLightkeeperRequired = task.lightkeeperRequired === true;
-      const isNonSpecial = !isKappaRequired && !isLightkeeperRequired;
-      if (hasTypeSelection) {
-        const matchesTaskType =
-          (isKappaRequired && showKappa) ||
-          (isLightkeeperRequired && showLightkeeper) ||
-          (isNonSpecial && showNonSpecial);
-        if (!matchesTaskType) continue;
-      }
-      if (onlyTasksWithRequiredKeys && !taskHasRequiredKeys(task)) {
-        continue;
-      }
-      const traderId = task.trader?.id;
-      if (!traderId) continue;
-      if (!counts[traderId]) counts[traderId] = 0;
-      const taskFaction = task.factionName;
-      if (isAllUsersView(userView)) {
-        const teamIds = Object.keys(progressStore.visibleTeamStores || {});
-        const relevantTeamIds = teamIds.filter((teamId) => {
-          const teamFaction = progressStore.playerFaction[teamId];
-          return taskFaction === 'Any' || taskFaction === teamFaction;
-        });
-        if (relevantTeamIds.length === 0) continue;
-        const taskStatuses = relevantTeamIds.map((teamId) => ({
-          isUnlocked: progressStore.unlockedTasks?.[task.id]?.[teamId] === true,
-          isCompleted: progressStore.tasksCompletions?.[task.id]?.[teamId] === true,
-          isFailed: progressStore.tasksFailed?.[task.id]?.[teamId] === true,
-        }));
-        let shouldCount = false;
-        switch (secondaryView) {
-          case 'all':
-            shouldCount = true;
-            break;
-          case 'available':
-            shouldCount = taskStatuses.some(isAvailableStatus);
-            break;
-          case 'locked':
-            if (isTaskInvalid(task.id, 'all')) continue;
-            shouldCount =
-              !taskStatuses.some(isAvailableStatus) &&
-              !taskStatuses.every(({ isCompleted }) => isCompleted) &&
-              !taskStatuses.some(({ isFailed }) => isFailed);
-            break;
-          case 'completed':
-            shouldCount = taskStatuses.every(
-              ({ isCompleted, isFailed }) => isCompleted && !isFailed
-            );
-            break;
-          case 'failed':
-            shouldCount = taskStatuses.some(({ isFailed }) => isFailed);
-            break;
-        }
-        if (shouldCount) counts[traderId]++;
-      } else {
-        const userFaction = progressStore.playerFaction[userView];
-        if (taskFaction !== 'Any' && taskFaction !== userFaction) continue;
-        const isUnlocked = progressStore.unlockedTasks?.[task.id]?.[userView] === true;
-        const isCompleted = progressStore.tasksCompletions?.[task.id]?.[userView] === true;
-        const isFailed = progressStore.tasksFailed?.[task.id]?.[userView] === true;
-        let shouldCount = false;
-        switch (secondaryView) {
-          case 'all':
-            shouldCount = true;
-            break;
-          case 'available':
-            shouldCount = isUnlocked && !isCompleted && !isFailed;
-            break;
-          case 'locked':
-            if (isTaskInvalid(task.id, userView)) continue;
-            shouldCount = !isCompleted && !isFailed && !isUnlocked;
-            break;
-          case 'completed':
-            shouldCount = isCompleted && !isFailed;
-            break;
-          case 'failed':
-            shouldCount = isFailed;
-            break;
-        }
-        if (shouldCount) counts[traderId]++;
-      }
-    }
-    perfEnd(perfTimer, { traders: Object.keys(counts).length });
-    return counts;
-  };
+  const { calculateStatusCounts, calculateTraderCounts } = useTaskCounts();
   return {
     visibleTasks,
     reloadingTasks,
